pyproject.toml
[project]
name = "mcp-server-evefleet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp==3.12.15",
    "anyio>=4.0.0",
    "certifi>=2024.12.14",
    "charset-normalizer>=3.0.0",
    "ecdsa==0.19.0",
    "idna>=3.7",
    "mcp[cli]>=1.14.0",
    "pyasn1==0.4.8",
    "python-jose==3.3.0",
    "pyyaml==6.0.1",
    "requests==2.32.3",
    "rsa==4.9",
    "six==1.16.0",
    "sniffio>=1.0.0",
    "urllib3>=2.2.0",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

